Feat(#31): OpenweatherClient 기능 구현
Resolvers: #31
